requirements.txt
boto3==1.21.3
botocore==1.24.3
Django==3.2.12
djangorestframework==3.13.1
nltk==3.7
numpy==1.22.2
pandas==1.4.1
protobuf==3.19.4
psycopg2-binary==2.9.3
sagemaker==1.72.0
scikit-learn==1.0.2
stanza==1.3.0
